fix(blade): rename `getPlatform` to `getPlatformType`
