--- packages/blade/src/utils/getPlatform/getPlatform.native.test.ts
@@ -1,7 +1,0 @@
-import getPlatform from './getPlatform';
-
-describe('getPlatform', () => {
-  it(`should return 'react-native'`, () => {
-    expect(getPlatform()).toEqual('react-native');
-  });
-});
