--- packages/blade/src/utils/getPlatform/index.ts
@@ -1,2 +1,0 @@
-export { default } from './getPlatform';
-export * from './getPlatform';
