+++ packages/blade/src/utils/getPlatformType/getPlatformType.native.test.ts
@@ -1,0 +1,7 @@
+import getPlatformType from './getPlatformType';
+
+describe('getPlatformType', () => {
+  it(`should return 'react-native'`, () => {
+    expect(getPlatformType()).toEqual('react-native');
+  });
+});
