+++ packages/blade/src/utils/getPlatformType/getPlatformType.web.test.ts
@@ -1,0 +1,12 @@
+import getPlatformType from './getPlatformType';
+
+describe('getPlatformType', () => {
+  it(`should return 'browser'`, () => {
+    expect(getPlatformType()).toEqual('browser');
+  });
+
+  it(`should return 'node'`, () => {
+    Object.defineProperty(global, 'document', {});
+    expect(getPlatformType()).toEqual('node');
+  });
+});
