+++ packages/blade/src/utils/getPlatformType/index.ts
@@ -1,0 +1,2 @@
+export { default } from './getPlatformType';
+export * from './getPlatformType';
